try again

+++ readout/readout.c
@@ -452,7 +452,7 @@ void *SensorThread(void *input){
             burstnum = ((unsigned int)slotval[slot] >> 24);
             voltage = (float)((float)adcdata / (float)(0x7FFF)) * 10;
 
-            int voltage_int = (int)voltage * 100;
+            int voltage_int = voltage * 100;
 
             printf('Voltage: V=%f, mV=%i \n', voltage, voltage_int);
             // update buffer
